Separated login page to new file, navigation now working!
