+++ iFeel-App/App.js
@@ -1,140 +1,12 @@
-import React from 'react';
-import { ActivityIndicator, StyleSheet, Text, View } from 'react-native';
-import * as firebase from 'firebase';
-import { Input } from './components/Input';
-import { Button } from './components/Button';
-
-export default class App extends React.Component {
-  state = {
-    email: '',
-    password: '',
-    authenticating: false,
-    user: null,
-    error: '',
-  }
-
-  // Initialize Firebase
-    componentWillMount() {
-    var config = {
-        apiKey: "",
-        authDomain: "ifeel-d97fc.firebaseapp.com",
-        databaseURL: "https://ifeel-d97fc.firebaseio.com",
-        projectId: "ifeel-d97fc",
-        storageBucket: "ifeel-d97fc.appspot.com",
-        messagingSenderId: "639485736592"
-        }
-
-        firebase.initializeApp(config);
-    }
-
-  onPressSignIn() {
-    this.setState({
-      authenticating: true,
-    });
-
-    const { email, password } = this.state;
-
-    firebase.auth().signInWithEmailAndPassword(email, password)
-      .then(user => this.setState({
-        authenticating: false,
-        user,
-        error: '',
-      }))
-      .catch(() => this.setState({
-            authenticating: false,
-            user: null,
-            error: 'Authentication Failure',
-          }))
-/*
-      .catch(() => {
-        // Login was not successful
-        firebase.auth().createUserWithEmailAndPassword(email, password)
-          .then(user => this.setState({
-            authenticating: false,
-            user,
-            error: '',
-          }))
-          .catch(() => this.setState({
-            authenticating: false,
-            user: null,
-            error: 'Authentication Failure',
-          }))
-      })
-      */
-  }
-  onPressLogOut() {
-    firebase.auth().signOut()
-      .then(() => {
-        this.setState({
-          email: '',
-          password: '',
-          authenticating: false,
-          user: null,
-        })
-      }, error => {
-        console.error('Sign Out Error', error);
-      });
-  }
-
-  renderCurrentState() {
-    if (this.state.authenticating) {
-      return (
-        <View style={styles.form}>
-          <ActivityIndicator size='large' />
-        </View>
-      )
-    }
-
-    if (this.state.user !== null) {
-      return (
-        <View style={styles.form}>
-          <Text>Logged In</Text>
-          <Button onPress={() => this.onPressLogOut()}>Let me out!</Button>
-        </View>
-      )
-    }
-
-    return (
-      <View style={styles.form}>
-        <Input
-          placeholder='Enter your email...'
-          label='Email'
-          onChangeText={email => this.setState({ email })}
-          value={this.state.email}
-        />
-        <Input
-          placeholder='Enter your password...'
-          label='Password'
-          secureTextEntry
-          onChangeText={password => this.setState({ password })}
-          value={this.state.password}
-        />
-        <Button onPress={() => this.onPressSignIn()}>Let me in!</Button>
-        <Text>{this.state.error}</Text>
-      </View>
-    )
-
-  }
-
-  render() {
-    return (
-      <View style={styles.container}>
-        {this.renderCurrentState()}
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 20,
-    alignItems: 'center',
-    justifyContent: 'center',
-    flexDirection: 'row',
-    backgroundColor: '#E84A27',
-  },
-  form: {
-    flex: 1
-  }
+// Import the screens
+import Main from './components/Main';
+import Chat from './components/Chat';
+// Import React Navigation
+import { createStackNavigator } from 'react-navigation'
+// Create the navigator
+const navigator = createStackNavigator({
+  Main: { screen: Main },
+  Chat: { screen: Chat },
 });
+// Export it as the root component
+export default navigator
