+++ iFeel-App/components/Chat.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+import { View, StyleSheet } from 'react-native';
+class Chat extends Component {
+  render() {
+    return <View />;
+  }
+}
+const styles = StyleSheet.create({});
+export default Chat;
