+++ iFeel-App/package.json
@@ -13,6 +13,7 @@
     "expo": "^30.0.1",
     "firebase": "^5.5.7",
     "react": "16.3.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz"
+    "react-native": "https://github.com/expo/react-native/archive/sdk-30.0.0.tar.gz",
+    "react-navigation": "^2.18.2"
   }
 }
